+++ src/bam/mod.rs
@@ -856,13 +856,15 @@ CCCCCCCCCCCCCCCCCCC"[..],
         let del_len = [1, 1, 1, 1, 1, 100000];
 
         for (i, record) in bam.records().enumerate() {
-            let rec = record.ok().expect("Expected valid record");
+            let mut rec = record.ok().expect("Expected valid record");
             println!("{}", str::from_utf8(rec.qname()).ok().unwrap());
             assert_eq!(rec.qname(), names[i]);
             assert_eq!(rec.flags(), flags[i]);
             assert_eq!(rec.seq().as_bytes(), seqs[i]);
-            assert_eq!(*rec.cigar(), cigars[i]);
-            let cigar = rec.cigar();
+            rec.unpack_cigar();
+            let cigar = rec.cigar().unwrap();
+            assert_eq!(**cigar, cigars[i]);
+
             if let Ok(end_pos) = cigar.end_pos() {
                 assert_eq!(end_pos, rec.pos() + 100 + del_len[i]);
                 assert_eq!(
@@ -958,12 +960,13 @@ CCCCCCCCCCCCCCCCCCC"[..],
             .ok()
             .expect("Expected successful fetch.");
         for (i, record) in bam.records().enumerate() {
-            let rec = record.ok().expect("Expected valid record");
+            let mut rec = record.ok().expect("Expected valid record");
+            rec.unpack_cigar();
             println!("{}", str::from_utf8(rec.qname()).ok().unwrap());
             assert_eq!(rec.qname(), names[i]);
             assert_eq!(rec.flags(), flags[i]);
             assert_eq!(rec.seq().as_bytes(), seqs[i]);
-            assert_eq!(*rec.cigar(), cigars[i]);
+            assert_eq!(**rec.cigar().unwrap(), cigars[i]);
             // fix qual offset
             let qual: Vec<u8> = quals[i].iter().map(|&q| q - 33).collect();
             assert_eq!(rec.qual(), &qual[..]);
@@ -985,9 +988,10 @@ CCCCCCCCCCCCCCCCCCC"[..],
         // note: this segfaults if you push_aux() before set()
         //       because set() obliterates aux
         rec.push_aux(b"NM", &Aux::Integer(15)).unwrap();
+        rec.unpack_cigar();
 
         assert_eq!(rec.qname(), names[0]);
-        assert_eq!(*rec.cigar(), cigars[0]);
+        assert_eq!(**rec.cigar().unwrap(), cigars[0]);
         assert_eq!(rec.seq().as_bytes(), seqs[0]);
         assert_eq!(rec.qual(), quals[0]);
         assert!(rec.is_reverse());
@@ -1003,9 +1007,10 @@ CCCCCCCCCCCCCCCCCCC"[..],
         let mut rec = record::Record::new();
         rec.set(names[0], &cigars[0], seqs[0], quals[0]);
         rec.push_aux(b"NM", &Aux::Integer(15)).unwrap();
+        rec.unpack_cigar();
 
         assert_eq!(rec.qname(), names[0]);
-        assert_eq!(*rec.cigar(), cigars[0]);
+        assert_eq!(**rec.cigar().unwrap(), cigars[0]);
         assert_eq!(rec.seq().as_bytes(), seqs[0]);
         assert_eq!(rec.qual(), quals[0]);
         assert_eq!(rec.aux(b"NM").unwrap(), Aux::Integer(15));
@@ -1020,9 +1025,10 @@ CCCCCCCCCCCCCCCCCCC"[..],
         let extension = b"BuffaloBUffaloBUFFaloBUFFAloBUFFALoBUFFALO";
         longer_name.extend(extension.iter());
         rec.set_qname(&longer_name);
+        rec.unpack_cigar();
 
         assert_eq!(rec.qname(), longer_name.as_slice());
-        assert_eq!(*rec.cigar(), cigars[0]);
+        assert_eq!(**rec.cigar().unwrap(), cigars[0]);
         assert_eq!(rec.seq().as_bytes(), seqs[0]);
         assert_eq!(rec.qual(), quals[0]);
         assert_eq!(rec.aux(b"NM").unwrap(), Aux::Integer(15));
@@ -1032,7 +1038,7 @@ CCCCCCCCCCCCCCCCCCC"[..],
         rec.set_qname(shorter_name);
 
         assert_eq!(rec.qname(), shorter_name);
-        assert_eq!(*rec.cigar(), cigars[0]);
+        assert_eq!(**rec.cigar().unwrap(), cigars[0]);
         assert_eq!(rec.seq().as_bytes(), seqs[0]);
         assert_eq!(rec.qual(), quals[0]);
         assert_eq!(rec.aux(b"NM").unwrap(), Aux::Integer(15));
@@ -1041,7 +1047,7 @@ CCCCCCCCCCCCCCCCCCC"[..],
         rec.set_qname(b"");
 
         assert_eq!(rec.qname(), b"");
-        assert_eq!(*rec.cigar(), cigars[0]);
+        assert_eq!(**rec.cigar().unwrap(), cigars[0]);
         assert_eq!(rec.seq().as_bytes(), seqs[0]);
         assert_eq!(rec.qual(), quals[0]);
         assert_eq!(rec.aux(b"NM").unwrap(), Aux::Integer(15));
@@ -1108,9 +1114,10 @@ CCCCCCCCCCCCCCCCCCC"[..],
             for i in 0..names.len() {
                 let mut rec = record::Record::new();
                 bam.read(&mut rec).ok().expect("Failed to read record.");
+                rec.unpack_cigar();
 
                 assert_eq!(rec.qname(), names[i]);
-                assert_eq!(*rec.cigar(), cigars[i]);
+                assert_eq!(**rec.cigar().unwrap(), cigars[i]);
                 assert_eq!(rec.seq().as_bytes(), seqs[i]);
                 assert_eq!(rec.qual(), quals[i]);
                 assert_eq!(rec.aux(b"NM").unwrap(), Aux::Integer(15));
@@ -1160,11 +1167,12 @@ CCCCCCCCCCCCCCCCCCC"[..],
             for (i, _rec) in bam.records().enumerate() {
                 let idx = i % names.len();
 
-                let rec = _rec.expect("Failed to read record.");
+                let mut rec = _rec.expect("Failed to read record.");
+                rec.unpack_cigar();
 
                 assert_eq!(rec.pos(), i as i32);
                 assert_eq!(rec.qname(), names[idx]);
-                assert_eq!(*rec.cigar(), cigars[idx]);
+                assert_eq!(**rec.cigar().unwrap(), cigars[idx]);
                 assert_eq!(rec.seq().as_bytes(), seqs[idx]);
                 assert_eq!(rec.qual(), quals[idx]);
                 assert_eq!(rec.aux(b"NM").unwrap(), Aux::Integer(15));