+++ src/bam/mod.rs
@@ -147,6 +147,26 @@ impl Reader {
         let _self = unsafe { &*(data as *mut Self) };
         unsafe { htslib::bam_read1(_self.bgzf, record) }
     }
+
+    /// Iterator over the records between the (optional) virtual offsets `start` and `end`
+    ///
+    /// # Arguments
+    ///
+    /// * `start` - Optional starting virtual offset to seek to. Throws an error if it is not
+    /// a valid virtual offset.
+    ///
+    /// * `end` - Read until the virtual offset is less than `end`
+    pub fn iter_chunk(&mut self, start: Option<i64>, end: Option<i64>) -> ChunkIterator<Self> {
+        if let Some(pos) = start {
+            self.seek(pos)
+                .expect("Failed to seek to the starting position");
+        };
+
+        ChunkIterator {
+            reader: self,
+            end: end,
+        }
+    }
 }
 
 impl Read for Reader {
@@ -452,6 +472,29 @@ impl<'a, R: Read> Iterator for Records<'a, R> {
     type Item = Result<record::Record, ReadError>;
 
     fn next(&mut self) -> Option<Result<record::Record, ReadError>> {
+        let mut record = record::Record::new();
+        match self.reader.read(&mut record) {
+            Err(ReadError::NoMoreRecord) => None,
+            Ok(()) => Some(Ok(record)),
+            Err(err) => Some(Err(err)),
+        }
+    }
+}
+
+/// Iterator over the records of a BAM until the virtual offset is less than `end`
+pub struct ChunkIterator<'a, R: 'a + Read> {
+    reader: &'a mut R,
+    end: Option<i64>,
+}
+
+impl<'a, R: Read> Iterator for ChunkIterator<'a, R> {
+    type Item = Result<record::Record, ReadError>;
+    fn next(&mut self) -> Option<Result<record::Record, ReadError>> {
+        if let Some(pos) = self.end {
+            if self.reader.tell() >= pos {
+                return None;
+            }
+        }
         let mut record = record::Record::new();
         match self.reader.read(&mut record) {
             Err(ReadError::NoMoreRecord) => None,
@@ -824,13 +867,15 @@ CCCCCCCCCCCCCCCCCCC"[..],
         let del_len = [1, 1, 1, 1, 1, 100000];
 
         for (i, record) in bam.records().enumerate() {
-            let rec = record.ok().expect("Expected valid record");
+            let mut rec = record.ok().expect("Expected valid record");
             println!("{}", str::from_utf8(rec.qname()).ok().unwrap());
             assert_eq!(rec.qname(), names[i]);
             assert_eq!(rec.flags(), flags[i]);
             assert_eq!(rec.seq().as_bytes(), seqs[i]);
-            assert_eq!(*rec.cigar(), cigars[i]);
-            let cigar = rec.cigar();
+            rec.unpack_cigar();
+            let cigar = rec.cigar().unwrap();
+            assert_eq!(**cigar, cigars[i]);
+
             if let Ok(end_pos) = cigar.end_pos() {
                 assert_eq!(end_pos, rec.pos() + 100 + del_len[i]);
                 assert_eq!(
@@ -926,12 +971,13 @@ CCCCCCCCCCCCCCCCCCC"[..],
             .ok()
             .expect("Expected successful fetch.");
         for (i, record) in bam.records().enumerate() {
-            let rec = record.ok().expect("Expected valid record");
+            let mut rec = record.ok().expect("Expected valid record");
+            rec.unpack_cigar();
             println!("{}", str::from_utf8(rec.qname()).ok().unwrap());
             assert_eq!(rec.qname(), names[i]);
             assert_eq!(rec.flags(), flags[i]);
             assert_eq!(rec.seq().as_bytes(), seqs[i]);
-            assert_eq!(*rec.cigar(), cigars[i]);
+            assert_eq!(**rec.cigar().unwrap(), cigars[i]);
             // fix qual offset
             let qual: Vec<u8> = quals[i].iter().map(|&q| q - 33).collect();
             assert_eq!(rec.qual(), &qual[..]);
@@ -953,9 +999,10 @@ CCCCCCCCCCCCCCCCCCC"[..],
         // note: this segfaults if you push_aux() before set()
         //       because set() obliterates aux
         rec.push_aux(b"NM", &Aux::Integer(15)).unwrap();
+        rec.unpack_cigar();
 
         assert_eq!(rec.qname(), names[0]);
-        assert_eq!(*rec.cigar(), cigars[0]);
+        assert_eq!(**rec.cigar().unwrap(), cigars[0]);
         assert_eq!(rec.seq().as_bytes(), seqs[0]);
         assert_eq!(rec.qual(), quals[0]);
         assert!(rec.is_reverse());
@@ -971,9 +1018,10 @@ CCCCCCCCCCCCCCCCCCC"[..],
         let mut rec = record::Record::new();
         rec.set(names[0], &cigars[0], seqs[0], quals[0]);
         rec.push_aux(b"NM", &Aux::Integer(15)).unwrap();
+        rec.unpack_cigar();
 
         assert_eq!(rec.qname(), names[0]);
-        assert_eq!(*rec.cigar(), cigars[0]);
+        assert_eq!(**rec.cigar().unwrap(), cigars[0]);
         assert_eq!(rec.seq().as_bytes(), seqs[0]);
         assert_eq!(rec.qual(), quals[0]);
         assert_eq!(rec.aux(b"NM").unwrap(), Aux::Integer(15));
@@ -988,9 +1036,10 @@ CCCCCCCCCCCCCCCCCCC"[..],
         let extension = b"BuffaloBUffaloBUFFaloBUFFAloBUFFALoBUFFALO";
         longer_name.extend(extension.iter());
         rec.set_qname(&longer_name);
+        rec.unpack_cigar();
 
         assert_eq!(rec.qname(), longer_name.as_slice());
-        assert_eq!(*rec.cigar(), cigars[0]);
+        assert_eq!(**rec.cigar().unwrap(), cigars[0]);
         assert_eq!(rec.seq().as_bytes(), seqs[0]);
         assert_eq!(rec.qual(), quals[0]);
         assert_eq!(rec.aux(b"NM").unwrap(), Aux::Integer(15));
@@ -1000,7 +1049,7 @@ CCCCCCCCCCCCCCCCCCC"[..],
         rec.set_qname(shorter_name);
 
         assert_eq!(rec.qname(), shorter_name);
-        assert_eq!(*rec.cigar(), cigars[0]);
+        assert_eq!(**rec.cigar().unwrap(), cigars[0]);
         assert_eq!(rec.seq().as_bytes(), seqs[0]);
         assert_eq!(rec.qual(), quals[0]);
         assert_eq!(rec.aux(b"NM").unwrap(), Aux::Integer(15));
@@ -1009,7 +1058,7 @@ CCCCCCCCCCCCCCCCCCC"[..],
         rec.set_qname(b"");
 
         assert_eq!(rec.qname(), b"");
-        assert_eq!(*rec.cigar(), cigars[0]);
+        assert_eq!(**rec.cigar().unwrap(), cigars[0]);
         assert_eq!(rec.seq().as_bytes(), seqs[0]);
         assert_eq!(rec.qual(), quals[0]);
         assert_eq!(rec.aux(b"NM").unwrap(), Aux::Integer(15));
@@ -1076,9 +1125,10 @@ CCCCCCCCCCCCCCCCCCC"[..],
             for i in 0..names.len() {
                 let mut rec = record::Record::new();
                 bam.read(&mut rec).ok().expect("Failed to read record.");
+                rec.unpack_cigar();
 
                 assert_eq!(rec.qname(), names[i]);
-                assert_eq!(*rec.cigar(), cigars[i]);
+                assert_eq!(**rec.cigar().unwrap(), cigars[i]);
                 assert_eq!(rec.seq().as_bytes(), seqs[i]);
                 assert_eq!(rec.qual(), quals[i]);
                 assert_eq!(rec.aux(b"NM").unwrap(), Aux::Integer(15));
@@ -1128,11 +1178,12 @@ CCCCCCCCCCCCCCCCCCC"[..],
             for (i, _rec) in bam.records().enumerate() {
                 let idx = i % names.len();
 
-                let rec = _rec.expect("Failed to read record.");
+                let mut rec = _rec.expect("Failed to read record.");
+                rec.unpack_cigar();
 
                 assert_eq!(rec.pos(), i as i32);
                 assert_eq!(rec.qname(), names[idx]);
-                assert_eq!(*rec.cigar(), cigars[idx]);
+                assert_eq!(**rec.cigar().unwrap(), cigars[idx]);
                 assert_eq!(rec.seq().as_bytes(), seqs[idx]);
                 assert_eq!(rec.qual(), quals[idx]);
                 assert_eq!(rec.aux(b"NM").unwrap(), Aux::Integer(15));