+++ CHANGELOG.md
@@ -8,8 +8,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.20.0] - 2018-06-18
 ### Added
+- Initial implementation of synced BCF reader interface.
 - Several small helper methods for BAM readers.
 ### Changes
+- Not skipping `fileformat=` header any more.
 - BCF records are always unpacked when reading.
 
 ## [0.19.1] - 2018-06-07